requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning Utils
scikit-learn>=1.3.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Progress Bars and Logging
tqdm>=4.65.0

# Experiment Tracking (optional but recommended)
tensorboard>=2.13.0

# Data Augmentation
albumentations>=1.3.0

# Audio/Music
pygame>=2.5.0
